refactor(app): move Labware page
