+++ app/src/App/DesktopApp.tsx
@@ -24,7 +24,7 @@ import { RobotSettings } from '../pages/Desktop/Devices/RobotSettings'
 import { ProtocolsLanding } from '../pages/Desktop/Protocols/ProtocolsLanding'
 import { ProtocolDetails } from '../pages/Desktop/Protocols/ProtocolDetails'
 import { AppSettings } from '../pages/AppSettings'
-import { Labware } from '../pages/Labware'
+import { Labware } from '../pages/Desktop/Labware'
 import { useSoftwareUpdatePoll } from './hooks'
 import { Navbar } from './Navbar'
 import { EstopTakeover, EmergencyStopContext } from '../organisms/EmergencyStop'
