+++ app/src/organisms/LabwareCard/index.tsx
@@ -22,7 +22,7 @@ import {
 
 import { UNIVERSAL_FLAT_ADAPTER_X_DIMENSION } from '../LabwareDetails/Gallery'
 import { CustomLabwareOverflowMenu } from './CustomLabwareOverflowMenu'
-import type { LabwareDefAndDate } from '../../pages/Labware/hooks'
+import type { LabwareDefAndDate } from '../../pages/Desktop/Labware/hooks'
 
 export interface LabwareCardProps {
   labware: LabwareDefAndDate
