+++ app/src/organisms/LabwareDetails/helpers/labels.ts
@@ -2,7 +2,7 @@ import uniqBy from 'lodash/uniqBy'
 import type {
   LabwareWellGroupProperties,
   LabwareDefinition,
-} from '../../../pages/Labware/types'
+} from '../../../pages/Desktop/Labware/types'
 const WELL_TYPE_BY_CATEGORY = {
   tubeRack: 'tube',
   tipRack: 'tip',
