+++ app/src/pages/Desktop/Labware/__tests__/Labware.test.tsx
@@ -2,24 +2,24 @@ import * as React from 'react'
 import { vi, it, describe, expect, beforeEach, afterEach } from 'vitest'
 import { fireEvent, screen } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
 import {
   useTrackEvent,
   ANALYTICS_OPEN_LABWARE_CREATOR_FROM_BOTTOM_OF_LABWARE_LIBRARY_LIST,
-} from '../../../redux/analytics'
-import { LabwareCard } from '../../../organisms/LabwareCard'
-import { AddCustomLabwareSlideout } from '../../../organisms/AddCustomLabwareSlideout'
-import { useToaster } from '../../../organisms/ToasterOven'
+} from '../../../../redux/analytics'
+import { LabwareCard } from '../../../../organisms/LabwareCard'
+import { AddCustomLabwareSlideout } from '../../../../organisms/AddCustomLabwareSlideout'
+import { useToaster } from '../../../../organisms/ToasterOven'
 import { useAllLabware, useLabwareFailure, useNewLabwareName } from '../hooks'
 import { Labware } from '..'
-import { mockDefinition } from '../../../redux/custom-labware/__fixtures__'
+import { mockDefinition } from '../../../../redux/custom-labware/__fixtures__'
 
-vi.mock('../../../organisms/LabwareCard')
-vi.mock('../../../organisms/AddCustomLabwareSlideout')
-vi.mock('../../../organisms/ToasterOven')
+vi.mock('../../../../organisms/LabwareCard')
+vi.mock('../../../../organisms/AddCustomLabwareSlideout')
+vi.mock('../../../../organisms/ToasterOven')
 vi.mock('../hooks')
-vi.mock('../../../redux/analytics')
+vi.mock('../../../../redux/analytics')
 
 const mockTrackEvent = vi.fn()
 const mockMakeSnackbar = vi.fn()
