+++ app/src/pages/Desktop/Labware/index.tsx
@@ -32,13 +32,13 @@ import { LabwareCreator } from '@opentrons/labware-library'
 import {
   useTrackEvent,
   ANALYTICS_OPEN_LABWARE_CREATOR_FROM_BOTTOM_OF_LABWARE_LIBRARY_LIST,
-} from '../../redux/analytics'
-import { addCustomLabwareFileFromCreator } from '../../redux/custom-labware'
-import { LabwareCard } from '../../organisms/LabwareCard'
-import { AddCustomLabwareSlideout } from '../../organisms/AddCustomLabwareSlideout'
-import { LabwareDetails } from '../../organisms/LabwareDetails'
-import { useToaster } from '../../organisms/ToasterOven'
-import { useFeatureFlag } from '../../redux/config'
+} from '../../../redux/analytics'
+import { addCustomLabwareFileFromCreator } from '../../../redux/custom-labware'
+import { LabwareCard } from '../../../organisms/LabwareCard'
+import { AddCustomLabwareSlideout } from '../../../organisms/AddCustomLabwareSlideout'
+import { LabwareDetails } from '../../../organisms/LabwareDetails'
+import { useToaster } from '../../../organisms/ToasterOven'
+import { useFeatureFlag } from '../../../redux/config'
 import { useAllLabware, useLabwareFailure, useNewLabwareName } from './hooks'
 
 import type { DropdownOption } from '@opentrons/components'
